# contracts/openapi/platform-draft.yaml
openapi: 3.1.0
info:
  title: OpsWeave platform API
  version: 0.1.0-draft
  description: >
    Identity-scoped business endpoints. Tenant and permissions are taken from the
    trusted principal, never from query or body. Explicit OIDC mode uses the authorization-code BFF;
    dev mode accepts only its configured loopback credential. Cookie requests require exact Origin
    and X-CSRF-TOKEN on unsafe methods; GET session metadata supplies the masked token.
    Zabbix Host ingest is available only when an explicit source mode is configured.
    All API responses use Cache-Control no-store and an X-OpsWeave-Request-Id UUID.
    The optional client request ID is correlation metadata only, not identity or idempotency.
paths:
...
  /api/v1/integrations/workflows:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
    get:
      operationId: pageWorkflows
      description: Read the owner-scoped workflow directory, operator catalog and bounded run summaries. The server derives tenant and permissions; no source is contacted.
      security: [{bearerAuth: []}, {browserSession: []}]
      responses:
        '200':
          description: Workflow directory
          content:
            application/json:
              schema: {$ref: '../schemas/v2/workflow-page.schema.json'}
        default: {description: 401 unauthenticated; 403 denied; 503 unavailable}
...
  /api/v1/integrations/workflows/host-schedules/workflows/{id}:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
      - {name: id, in: path, required: true, schema: {type: string, pattern: '^[a-z][a-z0-9_-]{0,47}$'}}
    get:
      operationId: getWorkflowHostSchedule
      description: Read the server-owned host schedule and task state for one published workflow. It is available only in the configured local or delegated runtime boundary.
      security: [{bearerAuth: []}, {browserSession: []}]
      responses:
        '200':
          description: Host schedule status
          content:
            application/json:
              schema: {$ref: '../schemas/v2/workflow-host-schedule-status.schema.json'}
        default: {description: 401 unauthenticated; 403 denied; 404 hidden workflow; 503 runtime unavailable}
...
  /api/v1/integrations/workflows/metric-streams/workflows/{id}:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
      - {name: id, in: path, required: true, schema: {type: string, pattern: '^[a-z][a-z0-9_-]{0,47}$'}}
      - {name: revision, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 10000}}
    get:
      operationId: getWorkflowMetricStream
      description: Read the bounded metric stream task and confirmed batch metadata. A revision query selects an immutable published version; it never reads upstream data.
      security: [{bearerAuth: []}, {browserSession: []}]
      responses:
        '200':
          description: Metric stream status
          content:
            application/json:
              schema: {$ref: '../schemas/v2/workflow-metric-stream-status.schema.json'}
        default: {description: 400 invalid revision; 401 unauthenticated; 403 denied; 404 hidden workflow; 503 stream unavailable}
  /api/v1/integrations/workflows/metric-streams/commands/{id}:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
      - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
    get:
      operationId: getWorkflowMetricStreamReceipt
      description: Read the original metric stream control receipt; it never repeats a command.
      security: [{bearerAuth: []}, {browserSession: []}]
      responses:
        '200':
          description: Metric stream receipt
          content:
            application/json:
              schema: {$ref: '../schemas/v2/workflow-metric-stream-receipt.schema.json'}
        default: {description: 401 unauthenticated; 403 denied; 404 hidden receipt; 503 stream unavailable}
...
  /api/v1/integrations/workflows/metric-streams/workflows/{id}/verification:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
      - {name: id, in: path, required: true, schema: {type: string, pattern: '^[a-z][a-z0-9_-]{0,47}$'}}
    post:
      operationId: verifyWorkflowMetricStream
      description: Confirm one bounded metric batch by id and return the refreshed stream status.
      security: [{bearerAuth: []}, {browserSession: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '../schemas/v2/workflow-metric-stream-verification.schema.json'}
      responses:
        '200':
          description: Updated metric stream status
          content:
            application/json:
              schema: {$ref: '../schemas/v2/workflow-metric-stream-status.schema.json'}
        default: {description: 400 invalid batch; 401 unauthenticated; 403 denied; 404 hidden batch; 409 already confirmed; 503 stream unavailable}
  /api/v1/integrations/workflows/log-streams/workflows/{id}:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
      - {name: id, in: path, required: true, schema: {type: string, pattern: '^[a-z][a-z0-9_-]{0,47}$'}}
      - {name: revision, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 10000}}
    get:
      operationId: getWorkflowLogStream
      description: Read bounded log stream task and confirmed batch metadata for one owner-scoped workflow version.
      security: [{bearerAuth: []}, {browserSession: []}]
      responses:
        '200':
          description: Log stream status
          content:
            application/json:
              schema: {$ref: '../schemas/v2/workflow-log-stream-status.schema.json'}
        default: {description: 400 invalid revision; 401 unauthenticated; 403 denied; 404 hidden workflow; 503 stream unavailable}
...
  /api/v1/integrations/workflows/quality/workflows/{id}/versions/{revision}:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
      - {name: id, in: path, required: true, schema: {type: string, pattern: '^[a-z][a-z0-9_-]{0,47}$'}}
      - {name: revision, in: path, required: true, schema: {type: integer, minimum: 1, maximum: 10000}}
    get:
      operationId: getWorkflowQualityReport
      description: Read bounded quality metadata and confirmed batch summaries for one immutable workflow version. Missing evidence remains explicit and no source or action is triggered.
      security: [{bearerAuth: []}, {browserSession: []}]
      responses:
        '200':
          description: Workflow quality report
          content:
            application/json:
              schema: {$ref: '../schemas/v2/workflow-quality-report.schema.json'}
        default: {description: 400 invalid path; 401 unauthenticated; 403 denied; 404 hidden version; 503 unavailable}
  /api/v1/integrations/workflows/quality/workflows/{id}/versions/{revision}/batches/{batchId}:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
      - {name: id, in: path, required: true, schema: {type: string, pattern: '^[a-z][a-z0-9_-]{0,47}$'}}
      - {name: revision, in: path, required: true, schema: {type: integer, minimum: 1, maximum: 10000}}
      - {name: batchId, in: path, required: true, schema: {type: string, format: uuid}}
    get:
      operationId: getWorkflowQualityBatch
      description: Read one owner-scoped quality batch summary with bounded coverage and reconciliation metadata.
      security: [{bearerAuth: []}, {browserSession: []}]
      responses:
        '200':
          description: Workflow quality batch
          content:
            application/json:
              schema: {$ref: '../schemas/v2/workflow-quality-batch.schema.json'}
        default: {description: 400 invalid path; 401 unauthenticated; 403 denied; 404 hidden batch; 503 unavailable}
  /api/v1/integrations/workflows/quality/workflows/{id}/versions/{revision}/diagnostics:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
      - {name: id, in: path, required: true, schema: {type: string, pattern: '^[a-z][a-z0-9_-]{0,47}$'}}
      - {name: revision, in: path, required: true, schema: {type: integer, minimum: 1, maximum: 10000}}
    get:
      operationId: getWorkflowDiagnosticReport
      description: Read bounded diagnostic observations derived from the workflow quality record. It does not call an external source or execute an action.
      security: [{bearerAuth: []}, {browserSession: []}]
      responses:
        '200':
          description: Diagnostic report
          content:
            application/json:
              schema: {$ref: '../schemas/v2/workflow-diagnostic-report.schema.json'}
        default: {description: 400 invalid path; 401 unauthenticated; 403 denied; 404 hidden version; 503 unavailable}
  /api/v1/integrations/workflows/quality/workflows/{id}/versions/{revision}/diagnostics/{observationId}:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
      - {name: id, in: path, required: true, schema: {type: string, pattern: '^[a-z][a-z0-9_-]{0,47}$'}}
      - {name: revision, in: path, required: true, schema: {type: integer, minimum: 1, maximum: 10000}}
      - {name: observationId, in: path, required: true, schema: {type: string, format: uuid}}
    get:
      operationId: getWorkflowDiagnosticObservation
      description: Read one immutable diagnostic observation and its closed result.
      security: [{bearerAuth: []}, {browserSession: []}]
      responses:
        '200':
          description: Diagnostic observation
          content:
            application/json:
              schema: {$ref: '../schemas/v2/workflow-diagnostic-observation.schema.json'}
        default: {description: 400 invalid path; 401 unauthenticated; 403 denied; 404 hidden observation; 503 unavailable}
...
components:
  parameters:
    clientRequestId:
      in: header
      name: X-OpsWeave-Request-Id
      required: false
      description: Optional single UUID; malformed or duplicate values return 400. Server generates one if absent and echoes a canonical UUID on every API response. Never grants permissions or suppresses duplicate writes.
      schema: {type: string, format: uuid}
      example: 6c8d4385-c206-4b59-8c2e-f9d34f8b3500
...
  securitySchemes:
    historyService:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Dedicated OAuth client_credentials access token profile described in ../history-service-identity.md. Never accepted as a browser or Runtime credential.
    runtimeKey:
      type: apiKey
      in: header
      name: X-OpsWeave-Runtime-Key
      description: Explicit local process attestation only; does not grant user permissions and is never exposed to Web/model.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Explicit loopback dev credential, or short-lived server-only Runtime delegation on allowlisted endpoints. Raw provider bearer JWTs are not accepted.
    browserSession:
      type: apiKey
      in: cookie
      name: __Host-opsweave
      description: OIDC BFF HttpOnly Secure SameSite=Lax host cookie. Explicit loopback protocol tests use opsweave-test-session. Unsafe methods also require X-CSRF-TOKEN and the configured Origin.